Split dnd data and event type files to one export per file
