+++ packages/twenty-front/src/modules/page-layout/components/PageLayoutTabListDroppableMoreButton.tsx
@@ -4,7 +4,7 @@ import { styled } from '@linaria/react';
 
 import { PAGE_LAYOUT_TAB_LIST_DROPPABLE_IDS } from '@/page-layout/components/PageLayoutTabListDroppableIds';
 import { PAGE_LAYOUT_TAB_DND_TYPE } from '@/page-layout/constants/PageLayoutTabDndType';
-import { type PageLayoutTabMoreButtonDropData } from '@/page-layout/types/PageLayoutWidgetDndData';
+import { type PageLayoutTabMoreButtonDropData } from '@/page-layout/types/PageLayoutTabMoreButtonDropData';
 import { TabMoreButton } from '@/ui/layout/tab-list/components/TabMoreButton';
 import { themeCssVariables } from 'twenty-ui/theme-constants';
