+++ packages/twenty-front/src/modules/page-layout/types/PageLayoutTabDragData.ts
@@ -1,0 +1,4 @@
+export type PageLayoutTabDragData = {
+  type: 'tab';
+  tabId: string;
+};
