+++ packages/twenty-front/src/modules/page-layout/types/PageLayoutTabListEndDropData.ts
@@ -1,0 +1,5 @@
+// beforeTabId null means append after the last tab.
+export type PageLayoutTabListEndDropData = {
+  type: 'tab-list-end';
+  beforeTabId: string | null;
+};
