+++ packages/twenty-front/src/modules/page-layout/types/PageLayoutTabMoreButtonDropData.ts
@@ -1,0 +1,3 @@
+export type PageLayoutTabMoreButtonDropData = {
+  type: 'tab-more-button';
+};
