+++ packages/twenty-front/src/modules/page-layout/types/PageLayoutTabWidgetDropData.ts
@@ -1,0 +1,4 @@
+export type PageLayoutTabWidgetDropData = {
+  type: 'tab-widget-drop';
+  tabId: string;
+};
