+++ packages/twenty-front/src/modules/page-layout/types/PageLayoutWidgetDragData.ts
@@ -1,0 +1,6 @@
+export type PageLayoutWidgetDragData = {
+  type: 'widget';
+  widgetId: string;
+  tabId: string;
+  index: number;
+};
