+++ packages/twenty-front/src/modules/page-layout/types/PageLayoutWidgetListDropData.ts
@@ -1,0 +1,4 @@
+export type PageLayoutWidgetListDropData = {
+  type: 'widget-list';
+  tabId: string;
+};
