+++ packages/twenty-front/src/modules/page-layout/widgets/fields/types/FieldsConfigurationFieldDragData.ts
@@ -1,0 +1,5 @@
+export type FieldsConfigurationFieldDragData = {
+  type: 'field';
+  groupId: string;
+  index: number;
+};
