+++ packages/twenty-front/src/modules/page-layout/widgets/fields/types/FieldsConfigurationFieldListEndDropData.ts
@@ -1,0 +1,5 @@
+// Catches drops below the last field of a list and drops into an empty group.
+export type FieldsConfigurationFieldListEndDropData = {
+  type: 'field-list-end';
+  groupId: string;
+};
