+++ packages/twenty-front/src/modules/page-layout/widgets/fields/types/FieldsConfigurationGroupDragData.ts
@@ -1,0 +1,5 @@
+export type FieldsConfigurationGroupDragData = {
+  type: 'group';
+  groupId: string;
+  index: number;
+};
