+++ packages/twenty-front/src/modules/page-layout/widgets/fields/types/FieldsConfigurationGroupListEndDropData.ts
@@ -1,0 +1,4 @@
+// Catches drops below the last group to append the dragged group at the end.
+export type FieldsConfigurationGroupListEndDropData = {
+  type: 'group-list-end';
+};
