+++ packages/twenty-front/src/modules/ui/utilities/drag-and-drop/types/DragDropProviderDragEndEvent.ts
@@ -1,0 +1,7 @@
+import { type Data } from '@dnd-kit/abstract';
+
+import { type DragDropProviderProps } from '@/ui/utilities/drag-and-drop/types/DragDropProviderProps';
+
+export type DragDropProviderDragEndEvent<TData extends Data> = Parameters<
+  NonNullable<DragDropProviderProps<TData>['onDragEnd']>
+>[0];
