+++ packages/twenty-front/src/modules/ui/utilities/drag-and-drop/types/DragDropProviderDragStartEvent.ts
@@ -1,0 +1,7 @@
+import { type Data } from '@dnd-kit/abstract';
+
+import { type DragDropProviderProps } from '@/ui/utilities/drag-and-drop/types/DragDropProviderProps';
+
+export type DragDropProviderDragStartEvent<TData extends Data> = Parameters<
+  NonNullable<DragDropProviderProps<TData>['onDragStart']>
+>[0];
