+++ packages/twenty-front/src/modules/ui/utilities/drag-and-drop/types/DragDropProviderProps.ts
@@ -1,0 +1,7 @@
+import { type Data } from '@dnd-kit/abstract';
+import { type DragDropProvider } from '@dnd-kit/react';
+import { type ComponentProps } from 'react';
+
+export type DragDropProviderProps<TData extends Data> = ComponentProps<
+  typeof DragDropProvider<TData>
+>;
